Agents/Master Agent/requirements.txt
# Core Google Cloud and AI libraries
google-adk==0.0.3
google-cloud-aiplatform>=1.38.0
google-auth>=2.20.0
google-api-core>=2.14.0

# Web framework (compatible versions)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
deprecated>=1.2.18

# Additional dependencies that may be required by google-adk
protobuf>=4.25.0
grpcio>=1.59.0